Part 4 bloglist expansion step 10

+++ part4/bloglist/controllers/blogs.js
@@ -3,13 +3,7 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -21,11 +15,9 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-
-    if (!token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
 
@@ -45,6 +37,16 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        if (decodedToken.id.toString() !== blog.user.toString()) {
+            return response.status(401).json({ error: 'unauthorized deletion attempt' })
+        }
+
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
     } catch (exception) {
